helpers and aesthetics
played around to learn about helpers. went over board.  Noticed that my syntax is not the same on all pages. Made a template to refer to to have a cohesive aesthetically structure. Have more views to finsh to be the same. going to try to add a groups id to post so post are connected via the group. then going to try to put post on the groups page.